+++ sakura_core/view/colors/CColor_Numeric.h
@@ -16,10 +16,10 @@ public:
 	CColor_Numeric() : m_nCOMMENTEND(0) { }
 	EColorIndexType GetStrategyColor() const override{ return COLORIDX_DIGIT; }
 	void InitStrategyStatus() override{ m_nCOMMENTEND = 0; }
-	bool BeginColor(const CStringRef& cStr, int nPos) override;
-	bool EndColor(const CStringRef& cStr, int nPos) override;
+	bool BeginColor(const CStringRef& cStr, ssize_t nPos) override;
+	bool EndColor(const CStringRef& cStr, ssize_t nPos) override;
 	bool Disp() const override{ return m_pTypeData->m_ColorInfoArr[COLORIDX_DIGIT].m_bDisp; }
 private:
-	int m_nCOMMENTEND;
+	ssize_t m_nCOMMENTEND;
 };
 #endif /* SAKURA_CCOLOR_NUMERIC_CC286AA9_0D00_4D7A_AF1A_F91018612171_H_ */
