+++ sakura_core/view/colors/CColor_Url.h
@@ -16,10 +16,10 @@ public:
 	CColor_Url() : m_nCOMMENTEND(0) { }
 	EColorIndexType GetStrategyColor() const override{ return COLORIDX_URL; }
 	void InitStrategyStatus() override{ m_nCOMMENTEND = 0; }
-	bool BeginColor(const CStringRef& cStr, int nPos) override;
-	bool EndColor(const CStringRef& cStr, int nPos) override;
+	bool BeginColor(const CStringRef& cStr, ssize_t nPos) override;
+	bool EndColor(const CStringRef& cStr, ssize_t nPos) override;
 	bool Disp() const override{ return m_pTypeData->m_ColorInfoArr[COLORIDX_URL].m_bDisp; }
 private:
-	int m_nCOMMENTEND;
+	ssize_t m_nCOMMENTEND;
 };
 #endif /* SAKURA_CCOLOR_URL_E9C938AC_E5C6_468E_925E_A7787635AB82_H_ */
